@@ -12,7 +12,7 @@ license = "MIT"
 "Release Notes" = "https://github.com/langchain-ai/langchain/releases?q=tag%3A%22langchain-cli%3D%3D0%22&expanded=true"
 
 [tool.poetry.dependencies]
-python = ">=3.8.1,<4.0"
+python = ">=3.9,<4.0"
 typer = { extras = ["all"], version = "^0.9.0" }
 gitpython = "^3.1.40"
 langserve = { extras = ["all"], version = ">=0.0.51" }
