@@ -37,6 +37,7 @@ from pydantic import (
     model_validator,
     validate_arguments,
 )
+from pydantic.v1 import BaseModel as BaseModelV1
 from typing_extensions import Annotated
 
 from langchain_core._api import deprecated
@@ -475,14 +476,27 @@ class ChildTool(BaseTool):
         if isinstance(tool_input, str):
             if input_args is not None:
                 key_ = next(iter(get_fields(input_args).keys()))
-                input_args.validate({key_: tool_input})
+                if hasattr(input_args, "model_validate"):
+                    input_args.model_validate({key_: tool_input})
+                else:
+                    input_args.parse_obj({key_: tool_input})
             return tool_input
         else:
             if input_args is not None:
-                result = input_args.parse_obj(tool_input)
+                if issubclass(input_args, BaseModel):
+                    result = input_args.model_validate(tool_input)
+                    result_dict = result.model_dump()
+                elif issubclass(input_args, BaseModelV1):
+                    result = input_args.parse_obj(tool_input)
+                    result_dict = result.dict()
+                else:
+                    raise NotImplementedError(
+                        "args_schema must be a Pydantic BaseModel, "
+                        f"got {self.args_schema}"
+                    )
                 return {
                     k: getattr(result, k)
-                    for k, v in result.dict().items()
+                    for k, v in result_dict.items()
                     if k in tool_input
                 }
             return tool_input
