@@ -106,7 +106,7 @@ def _get_configs_for_single_dir(job: str, dir_: str) -> List[Dict[str, str]]:
     if dir_ == "libs/core":
         return [
             {"working-directory": dir_, "python-version": f"3.{v}"}
-            for v in range(8, 13)
+            for v in range(9, 13)
         ]
     min_python = "3.9"
     max_python = "3.12"
