@@ -4,11 +4,12 @@ from __future__ import annotations
 
 import inspect
 import textwrap
+import warnings
 from functools import wraps
 from typing import Any, Callable, Dict, List, Optional, Type, TypeVar, Union, overload
 
 import pydantic
-from pydantic import BaseModel, root_validator
+from pydantic import BaseModel, PydanticDeprecationWarning, root_validator
 from pydantic.json_schema import GenerateJsonSchema, JsonSchemaValue
 from pydantic_core import core_schema
 
@@ -134,42 +135,45 @@ def pre_init(func: Callable) -> Any:
         Any: The decorated function.
     """
 
-    @root_validator(pre=True)
-    @wraps(func)
-    def wrapper(cls: Type[BaseModel], values: Dict[str, Any]) -> Dict[str, Any]:
-        """Decorator to run a function before model initialization.
+    with warnings.catch_warnings():
+        warnings.filterwarnings(action="ignore", category=PydanticDeprecationWarning)
 
-        Args:
-            cls (Type[BaseModel]): The model class.
-            values (Dict[str, Any]): The values to initialize the model with.
+        @root_validator(pre=True)
+        @wraps(func)
+        def wrapper(cls: Type[BaseModel], values: Dict[str, Any]) -> Dict[str, Any]:
+            """Decorator to run a function before model initialization.
 
-        Returns:
-            Dict[str, Any]: The values to initialize the model with.
-        """
-        # Insert default values
-        fields = cls.model_fields
-        for name, field_info in fields.items():
-            # Check if allow_population_by_field_name is enabled
-            # If yes, then set the field name to the alias
-            if hasattr(cls, "Config"):
-                if hasattr(cls.Config, "allow_population_by_field_name"):
-                    if cls.Config.allow_population_by_field_name:
+            Args:
+                cls (Type[BaseModel]): The model class.
+                values (Dict[str, Any]): The values to initialize the model with.
+
+            Returns:
+                Dict[str, Any]: The values to initialize the model with.
+            """
+            # Insert default values
+            fields = cls.model_fields
+            for name, field_info in fields.items():
+                # Check if allow_population_by_field_name is enabled
+                # If yes, then set the field name to the alias
+                if hasattr(cls, "Config"):
+                    if hasattr(cls.Config, "allow_population_by_field_name"):
+                        if cls.Config.allow_population_by_field_name:
+                            if field_info.alias in values:
+                                values[name] = values.pop(field_info.alias)
+                if hasattr(cls, "model_config"):
+                    if cls.model_config.get("populate_by_name"):
                         if field_info.alias in values:
                             values[name] = values.pop(field_info.alias)
-            if hasattr(cls, "model_config"):
-                if cls.model_config.get("populate_by_name"):
-                    if field_info.alias in values:
-                        values[name] = values.pop(field_info.alias)
 
-            if name not in values or values[name] is None:
-                if not field_info.is_required():
-                    if field_info.default_factory is not None:
-                        values[name] = field_info.default_factory()
-                    else:
-                        values[name] = field_info.default
+                if name not in values or values[name] is None:
+                    if not field_info.is_required():
+                        if field_info.default_factory is not None:
+                            values[name] = field_info.default_factory()
+                        else:
+                            values[name] = field_info.default
 
-        # Call the decorated function
-        return func(cls, values)
+            # Call the decorated function
+            return func(cls, values)
 
     return wrapper
 
@@ -184,34 +188,6 @@ class _IgnoreUnserializable(GenerateJsonSchema):
         self, schema: core_schema.CoreSchema, error_info: str
     ) -> JsonSchemaValue:
         return {}
-
-
-def v1_repr(obj: BaseModel) -> str:
-    """Return the schema of the object as a string.
-
-    Get a repr for the pydantic object which is consistent with pydantic.v1.
-    """
-    if not is_basemodel_instance(obj):
-        raise TypeError(f"Expected a pydantic BaseModel, got {type(obj)}")
-    repr_ = []
-    for name, field in get_fields(obj).items():
-        value = getattr(obj, name)
-
-        if isinstance(value, BaseModel):
-            repr_.append(f"{name}={v1_repr(value)}")
-        else:
-            if field.exclude:
-                continue
-            if not field.is_required():
-                if not value:
-                    continue
-                if field.default == value:
-                    continue
-
-            repr_.append(f"{name}={repr(value)}")
-
-    args = ", ".join(repr_)
-    return f"{obj.__class__.__name__}({args})"
 
 
 def _create_subset_model_v1(
