@@ -12,7 +12,6 @@ from langchain_core.utils.pydantic import (
     is_basemodel_instance,
     is_basemodel_subclass,
     pre_init,
-    v1_repr,
 )
 
 
@@ -172,7 +171,7 @@ def test_fields_pydantic_v1() -> None:
         x: int
 
     fields = get_fields(Foo)
-    assert fields == {"x": Foo.__fields__["x"]}  # type: ignore[index]
+    assert fields == {"x": Foo.model_fields["x"]}  # type: ignore[index]
 
 
 @pytest.mark.skipif(PYDANTIC_MAJOR_VERSION != 2, reason="Only tests Pydantic v2")
@@ -195,17 +194,3 @@ def test_fields_pydantic_v1_from_2() -> None:
 
     fields = get_fields(Foo)
     assert fields == {"x": Foo.__fields__["x"]}
-
-
-def test_v1_repr() -> None:
-    from pydantic import BaseModel, Field
-
-    class Foo(BaseModel):
-        bar: int = Field(default=1, alias="baz", exclude=True)
-        x: int = Field(default=2)
-
-        def __repr__(self) -> str:
-            """Custom repr."""
-            return v1_repr(self)
-
-    assert repr(Foo(bar=2, x=3)) == "Foo(x=3)"  # type: ignore[call-arg]
